Struct oxagaudiotool::audio_tool::OxAgAudioTool

pub struct OxAgAudioTool { /* private fields */ }

Struct used to play the various sounds

Implementations

impl OxAgAudioTool

pub fn new( event_to_sound_config: HashMap<Event, OxAgSoundConfig>, tile_type_to_sound_config: HashMap<TileType, OxAgSoundConfig>, weather_type_to_sound_config: HashMap<WeatherType, OxAgSoundConfig> ) -> Result<OxAgAudioTool, OxAgAudioToolError>

Parameters

event_to_sound_config - a HashMap that maps 0 or more Events to a given OxAgSoundConfig

tile_type_to_sound_config - a HashMap that maps 0 or more TileTypes to a given OxAgSoundConfig

weather_type_to_sound_config - a WeatherType that maps 0 or more TileTypes to a given OxAgSoundConfig

pub fn play_audio_based_on_event( &mut self, event: &Event ) -> Result<(), OxAgAudioToolError>

Processes an event and plays the appropriate audio for it

If it’s a Event::Moved it will detect whether the TileType on which the robot is on changed and if it did it plays the configured sound for it, stopping the one playing previously for the different TileType.

If it’s either a Event::TimeChanged or a Event::DayChanged it will detect whether the WeatherType has changed and if it did it plays the configured sound for it, stopping the one playing previously for the different WeatherType.

If it’s any other event, it checks whether a sound is configured for it and plays it

pub fn play_audio( &mut self, sound_config: &OxAgSoundConfig ) -> Result<(), OxAgAudioToolError>

Plays an audio given some OxAgSoundConfig