Struct oxagaudiotool::sound_config::OxAgSoundConfig

pub struct OxAgSoundConfig { /* private fields */ }

Configures any sound that should be played via this library

Properties

path - the path to the audio file, supported formats are: mp3, ogg, wav, flac

settings - the actual StaticSoundSettings

Implementations

impl OxAgSoundConfig

pub fn new(path: &str) -> OxAgSoundConfig

Params

path - the path to the audio file, supported formats are: mp3, ogg, wav, flac

pub fn new_with_settings( path: &str, settings: StaticSoundSettings ) -> OxAgSoundConfig

Params

path - the path to the audio file, supported formats are: mp3, ogg, wav, flac

settings - fully customisable sound settings from the kira library

pub fn new_looped(path: &str) -> OxAgSoundConfig

Params

path - the path to the audio file, supported formats are: mp3, ogg, wav, flac

Side effects

When the sound gets played it will get looped infinitely by default

pub fn new_with_volume(path: &str, amplitude: f64) -> OxAgSoundConfig

Params

path - the path to the audio file, supported formats are: mp3, ogg, wav, flac

amplitude - the sound will be played with the default volume multiplied by the provided amplitude

pub fn new_looped_with_volume(path: &str, amplitude: f64) -> OxAgSoundConfig

Params

path - the path to the audio file, supported formats are: mp3, ogg, wav, flac

amplitude - the sound will be played with the default volume multiplied by the provided amplitude

Side effects

When the sound gets played it will get looped infinitely by default

pub fn looped(&mut self)

When the sound gets played it will get looped infinitely by default

pub fn volume(&mut self, amplitude: f64)

The sound will be played with the default volume multiplied by the provided amplitude